12 Openai agentic SDK/xx_project/01_multiagents/multiagents/pyproject.toml
[project]
name = "multiagents"
version = "0.1.0"
description = "Multi-agent system using OpenAI's SDK"
readme = "README.md"
authors = [
    { name = "Maimuzamilhu", email = "muzamilkhalid808@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai",
    "python-dotenv",
    "agents"
]

[project.scripts]
multiagents = "multiagents:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
